tools/tools/ingestor/requirements.txt
httpx
curl_cffi
cloudscraper
requests
beautifulsoup4
lxml
pillow
python-dotenv
groq
supabase
boto3
pydantic
